finished rendering logic
Finished logic for making the markup for main-content, form, and chess board grid. Created a file called 'eventlistener.js' which has our code for creating and developing the eventlisteners for the application. Finally finished drawing logic for drawing the path for the chessboard. Now that the main logic for the application is done, we're going to be working on the css for things such as changing themes, making it responsive, etc.