Finished it seems like
Fixed, issue with drop down so that even when the user is resizing it will display correctly, though the animation was removed because of this. Removed eventlisteners.js because we could just put the main code in app.js rather than creating a whole other file. Set up logic so that we could render a knight path when the page loads, based on what the dev puts in the modules.js. Also for the drop down, the dev can control whether the page loads the menu collapsed or hidden on smaller screens with the menuCollapsed boolean in statesModule. Good project, and even though it isn't noticeable with the layout I used, i'm glad I designed with the idea of doing the mobile layout first.